N64 magic. Bizhawk's API expects certain methods, one of which is Frame_Advance()
Advance_Frame() is a generic method which will then call any core-specific stuff required to advance the frame.

In the case of Mupen64Plus, when calling the unmanaged frame advance function from within mupen, it uses a method named HackyPinvokeWaitOne which is essentially just a modified version of WaitHandle.WaitOne(), which blocks the thread until the WaitHandle receives a response.

In the case of Mupen64Plus crashing internally, that WaitHandle() never gets a response, and thus, the thread is blocked forever as the default timeout value is Int32.MaxValue

By changing the timeout to something reasonable like 1000ms (way longer than one frame should ever take), you can get it to no longer block the thread forever. Now this doesn't completely solve the problem as Bizhawk has no idea that the frame advance failed, and as such, keeps trying to call Frame_Advance(), blocking the thread for another 1000ms.

By adding a simple bool that's set after a timeout, we can tell it to stop calling the internal Frame_Advance() method which results in the thread not being blocked any more so you can reboot the core from the main thread.